fdfg

+++ controller/admin_controller.js
@@ -19,7 +19,6 @@ const allexpense = async (req, res) => {
     }
 }
 
-
 // *--------------------------------------
 // * Admin get all user Logic
 // *--------------------------------------
@@ -37,13 +36,17 @@ const alluser = async (req, res) => {
     }
 }
 
-
 // *--------------------------------------
-// * Admin get all user Logic
+// * Admin user data update Logic
 // *--------------------------------------
 const userupdate = async (req, res) => {
     // console.log(req.body);
-    const { id, name, phone, email, admin,verified } = req.body
+    const { id, name, phone, email, admin,verified } = req.body;
+    if(!(id, name, phone, email, admin,verified)){
+        return res.status(422).json({
+            msg: "All fields are required"
+        })
+    }
     try {
         const query = await user.findByIdAndUpdate({ _id: id }, { name, phone, email, isadmin: admin, isverified:verified });
         if (query) {
@@ -57,11 +60,16 @@ const userupdate = async (req, res) => {
 }
 
 // *--------------------------------------
-// * Admin get all user Logic
+// * Admin user delete & user Expense Delete Logic
 // *--------------------------------------
 const removeuser = async (req, res) => {
     // console.log(req.body);
     const { id } = req.body
+    if(!id){
+        return res.status(422).json({
+            msg: "ID is required"
+        })
+    }
     try {
         const query = await user.findByIdAndDelete({ _id: id });
         const exp = await expense.remove({userid:id});
